requirements.txt
streamlit
google-generativeai
python-dotenv

langchain
langchain-community
langchain-huggingface
langchain-text-splitters

sentence-transformers
faiss-cpu
pypdf

